organize pytorch files
